s3/twitter/requirements_kafka.txt
# Python Packages
boto3
tweepy
kafka-python